% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chf}
\alias{chf}
\title{Heart failure exacerbations}
\format{
\subsection{\code{chf}}{

A data frame with 334 rows and 4 columns
\describe{
\item{id}{Subject ID}
\item{adhf}{Acute decompensated heart failure event}
\item{date}{Date of event}
\item{lvef}{Left ventricular ejection fraction at time of event}
}
}
}
\usage{
chf
}
\description{
Heart failure exacerbations
}
\keyword{datasets}
