% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lzrs-package.R
\docType{package}
\name{lzrs-package}
\alias{lzrs}
\alias{lzrs-package}
\title{lzrs: Longitudinal and recurrent event statistics}
\description{
Tools to shape and analyze recurrent events
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/shiva}
  \item \url{https://asshah4.github.io/shiva/}
  \item Report bugs at \url{https://github.com/asshah4/shiva/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
