% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcr-package.R
\docType{package}
\name{rcr-package}
\alias{rcr}
\alias{rcr-package}
\title{rcr: Longitudinal and recurrent event statistics}
\description{
Tools to shape and analyze recurrent events
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/rcr}
  \item \url{https://asshah4.github.io/rcr/}
  \item Report bugs at \url{https://github.com/asshah4/rcr/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
